Move utils to the only service that uses them
